# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Oracle Factory server.",
        "title": "Oracle Factory API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/outboundOracles": {
            "get": {
                "description": "Retrieve all outbound oracles of a user. This will be called from the frontend, when a user wants to view a list of oracle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Retrieves all outbound oracle of a user",
                "responses": {}
            },
            "post": {
                "description": "Creates an outbound oracle for a user. This service will be called by the frontend to when a user filled out the outbound oracle form.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Creates an outbound oracle for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/outboundOracles/{outboundOracleID}": {
            "get": {
                "description": "Retrieve the specified outbound oracle. This will be called from the frontend, when a user wants to view an oracle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Retrieve an outbound oracle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the outbound oracle you want to retrieve.",
                        "name": "outboundOracleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Update the specified outbound oracle. This will be called from the frontend, when a user wants to update an oracle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Update an outbound oracle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the outbound oracle you want to update.",
                        "name": "outboundOracleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/outboundOracles/{outboundOracleID}/start": {
            "post": {
                "description": "Start the specified outbound oracle. This will be called from the frontend, when a user wants to use an oracle for a blockchain conenction.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Start an Outbound Oracle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the outbound oracle you want to start.",
                        "name": "outboundOracleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/outboundOracles/{outboundOracleID}/stop": {
            "post": {
                "description": "Stop the specified outbound oracle. This will be called from the frontend, when a user wants to stop an oracle for a blockchain conenction.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outboundOracles"
                ],
                "summary": "Stop an outbound oracle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the outbound oracle you want to stop.",
                        "name": "outboundOracleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "description": "Retrieve the signed in user. This will be called by the frontend to get all information about the user signed in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve signed in user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user. This will be called from the frontend to update the settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "responses.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}